Import fixes.

+++ src/main/kotlin/ru/tesserakt/diskordin/core/data/json/request/GuildRoleEditRequest.kt
@@ -3,7 +3,7 @@ package ru.tesserakt.diskordin.core.data.json.request
 
 data class GuildRoleEditRequest(
     val name: String? = null,
-    val permissions: Int? = null,
+    val permissions: Long? = null,
     val color: Int? = null,
     val hoist: Boolean? = null,
     val mentionable: Boolean? = null
