+++ src/main/kotlin/ru/tesserakt/diskordin/core/data/json/response/OverwriteResponse.kt
@@ -4,6 +4,6 @@ package ru.tesserakt.diskordin.core.data.json.response
 data class OverwriteResponse(
     val id: Long,
     val type: String,
-    val allow: Int,
-    val deny: Int
+    val allow: Long,
+    val deny: Long
 ) : DiscordResponse()
